+++ src/xSTIR/cSTIR/include/sirf/STIR/stir_types.h
@@ -63,6 +63,10 @@ limitations under the License.
 #include "stir/SSRB.h"
 #include "stir/TruncateToCylindricalFOVImageProcessor.h"
 
+#ifdef STIR_WITH_NIFTYPET_PROJECTOR
+#include "stir/recon_buildblock/niftypet_projector/ProjectorByBinPairUsingNiftyPET.h"
+#endif
+
 #include "stir/StirException.h"
 #include "stir/TextWriter.h"
 
@@ -85,6 +89,9 @@ namespace sirf {
 		PoissonLogLhLinModMean3DF;
 	//PoissonLogLikelihoodWithLinearModelForMeanAndProjData<Image3DF>
 	typedef stir::ProjectorByBinPairUsingProjMatrixByBin ProjectorPairUsingMatrix;
+#ifdef STIR_WITH_NIFTYPET_PROJECTOR
+    typedef stir::ProjectorByBinPairUsingNiftyPET ProjectorPairUsingNiftyPET;
+#endif
 	typedef stir::ProjMatrixByBinUsingRayTracing RayTracingMatrix;
 	typedef stir::GeneralisedPrior<Image3DF> Prior3DF;
 	typedef stir::QuadraticPrior<float> QuadPrior3DF;